secondo commit

+++ code/calcolaPotenza/src/calcolaPotenza/it/corso/driver.java
@@ -9,14 +9,14 @@ public class driver {
 	System.out.println("\nFOR");
 		for (i = 1; i <= 16; i++) {
 			
-			System.out.println("2 elevato a " + i + " = " + Math.pow(2, i));
+			System.out.println("2 elevato a " + (int)i + " = " + Math.pow(2, i));
 		}
 		
 		i = 1;
 		System.out.println("\nWHILE");
 		while (i<16) {
 			
-			System.out.println("2 elevato a " + i + " = " + Math.pow(2, i));
+			System.out.println("2 elevato a " + (int)i + " = " + Math.pow(2, i));
 			i++;
 		}
